moved comparison function to own file
