+++ server/imageCompare/index.js
@@ -1,7 +1,8 @@
-const compareImages = require('resemblejs/compareImages');
-const fs = require('fs');
+/// this is just a page to check the comparison function lcoated in getDiff.js
+/// we can delete this for production
 
 const express = require('express');
+const getDiff = require('./getDiff.js');
 
 const app = express();
 
@@ -12,20 +13,6 @@ const assetUrls = ['testAssets/black.jpg',
   'testAssets/topblack.jpg',
   'testAssets/bottomblack.jpg',
 ];
-
-async function getDiff(path1, path2) {
-  try {
-    const img1 = fs.readFileSync(path1);
-    const img2 = fs.readFileSync(path2);
-
-    const data = await compareImages(img1, img2);
-    console.log('DATA', data);
-    return data;
-  } catch (error) {
-    console.log('ERROR=========================', error);
-    return error;
-  }
-}
 
 app.get('/', (req, res) => {
   const promises = [];
